% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{calculate_flow_duration}
\alias{calculate_flow_duration}
\title{Calculate Flow Duration Curve
Generate flow duration curve data.}
\usage{
calculate_flow_duration(data)
}
\arguments{
\item{data}{Yarra River data frame}
}
\value{
Flow duration curve data
}
\description{
Calculate Flow Duration Curve
Generate flow duration curve data.
}
