% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YarraWaterAnalysis-package.R
\docType{package}
\name{YarraWaterAnalysis-package}
\alias{YarraWaterAnalysis}
\alias{YarraWaterAnalysis-package}
\title{YarraWaterAnalysis: Yarra River Streamflow Data Analysis and Visualization}
\description{
Provides tools for analyzing and visualizing Yarra River streamflow data from the McMahons monitoring site. Includes interactive Shiny applications for data exploration and various analytical functions for hydrological analysis.
}
\author{
\strong{Maintainer}: Chen Liu \email{cliu0197@student.monash.edu}

}
\keyword{internal}
